upgrade to .net 6.0 LTS;
1.upgrade to .net 6.0 LTS;
2.change comment status;
3.reply to comment;

+++ src/Lequ/Controllers/AboutController.cs
@@ -18,7 +18,7 @@ namespace Lequ.Controllers
         }
         public async Task<IActionResult> Index()
         {
-            var value = await _service.GetAsync(x => x.Status == ModelStatus.Normal);
+            var value = await _service.GetAsync(x => x.Status == (int)ModelStatus.Normal);
             return View(value);
         }
