+++ src/Lequ/ViewComponents/Comment/CommentList.cs
@@ -21,7 +21,7 @@ namespace Lequ.ViewComponents.Comment
 
         public async Task<IViewComponentResult> InvokeAsync(int blogID)
         {
-            var comments = await _service.SelectAsync(x => x.BlogID == blogID && x.Status == ModelStatus.Normal, x => x.Parent);
+            var comments = await _service.SelectAsync(x => x.BlogID == blogID && x.Status == (int)ModelStatus.Normal, x => x.Parent);
             var vm = new BlogCommentListViewModel();
             vm.BlogID = blogID;
             if (comments != null && comments.Count > 0)
